double

+++ kmeans.c
@@ -403,7 +403,7 @@ int main(int argc, char* argv[])
 	}
 
 	local_matrix = (float *)malloc(sendcnts[rank]*sizeof(double));
-    MPI_Scatterv(&data, sendcnts, displs, MPI_FLOAT, local_matrix, sendcnts[rank], MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    MPI_Scatterv(&data, sendcnts, displs, MPI_DOUBLE, local_matrix, sendcnts[rank], MPI_DOUBLE, 0, MPI_COMM_WORLD);
 	int count_info = K * samples;
 	MPI_Bcast(&centroids, count_info, MPI_FLOAT, 0, MPI_COMM_WORLD);
 	int info[3] = {K, lines, samples};
